[#3530] Moved roll5e changes to BasicRoll
Removed changes that added Roll5e (all previous commits from PR should be reversed)
Added Term replacement methods to BasicRoll
Modified damage-roll to create a BasicRoll and run evaluateSync with the maximize option.
Removed v11 checks for async options
Closes #3530
