+++ module/dice/_module.mjs
@@ -1,6 +1,5 @@
 export {default as aggregateDamageRolls} from "./aggregate-damage-rolls.mjs";
 export {default as BasicRoll} from "./basic-roll.mjs";
-export {default as Roll5e} from "./roll.mjs";
 export {default as D20Roll} from "./d20-roll.mjs";
 export {default as DamageRoll} from "./damage-roll.mjs";
 export * from "./dice.mjs";
